requirements.txt
gunicorn==22.0.0
Flask==3.0.0
black==24.3.0
mypy==1.15.0
pymysql==1.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
types-Flask==1.1.6
requests==2.32.5
types-requests==2.32.0.20241016
